Fix finding source files in subdirectories
This fixes issue with searching for source file if header is included
using relative path ex. #include "one/two/three.hpp".

The algorithm is searching for source files in following directories:
 - directory where main source file was located,
 - additional source directory (if it was specified).

But it was searching only for file with same basename as header. It was
skipping relative path used in include directive, so it was not able to
find source file if it was located in subdirectory.

Introduced fix is changing this logic. Algorithm is now always use
relative path when searching for corresponding source file for included
headers. For example, if header file is included using following
directive:

#include "one/two/three.hpp"

The source file "three.cpp" should be located in directory "one/two/"
or "${sourceDirectory}/one/two/".